fix checkbox + disabling menu and bench + cleaning index.js
