+++ src/error/index.js
@@ -1,21 +1,22 @@
-var Collection = require('./model');
-var View = require('./view');
+var model = require('./model');
+var view = require('./view');
 //
 module.exports = function (app) {
   app.on('before:start', function () {
-    var collection = new Collection();
-    collection.listenTo(app, 'error:add', function (error) {
+    var errors = new model.ErrorCollection();
+    app.getView()
+      .getRegion('error')
+      .show(new view.Errors({
+        collection: errors
+      }));
+    // Incoming
+    errors.listenTo(app, 'error:add', function (error) {
       this.add({
         message: 'Error : ' + error,
       });
     });
-    collection.listenTo(app, 'error:reset', function () {
+    errors.listenTo(app, 'error:reset', function () {
       this.reset();
     });
-    app.getView()
-      .getRegion('error')
-      .show(new View({
-        collection: collection
-      }));
   });
 };
